+++ js/tweek-client/src/TweekManagementClient/types.ts
@@ -1,0 +1,4 @@
+export interface ITweekManagementClient {
+  appendContext(identityType: string, identityId: string, context: object): Promise<void>;
+  deleteContext(identityType: string, identityId: string, property: string): Promise<void>;
+}
